# specs/apps/crud/contracts/generated/openapi-bundled.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "crud API",
    "description": "REST API contract for the polyglot crud expense tracker crud application",
    "version": "1.0.0",
    "contact": {
      "name": "ose-primer",
      "url": "https://github.com/wahidyankf/ose-primer"
    }
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Default backend (Go/Gin)"
    },
    {
      "url": "http://localhost:8201",
      "description": "Java Spring Boot backend"
    }
  ],
...
  "paths": {
...
    "/api/v1/auth/refresh": {
      "post": {
        "operationId": "refreshToken",
        "summary": "Refresh access token",
        "tags": [
          "Authentication"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RefreshRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Token refreshed successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthTokens"
                }
              }
            }
          },
          "401": {
            "description": "Invalid or expired refresh token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/auth/logout-all": {
      "post": {
        "operationId": "logoutAll",
        "summary": "Logout all sessions",
        "tags": [
          "Authentication"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "All sessions logged out"
          }
        }
      }
    },
...
    "/api/v1/expenses/{id}/attachments": {
      "post": {
        "operationId": "uploadAttachment",
        "summary": "Upload attachment to expense",
        "tags": [
          "Attachments"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "required": [
                  "file"
                ],
                "properties": {
                  "file": {
                    "type": "string",
                    "format": "binary"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Attachment uploaded",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Attachment"
                }
              }
            }
          },
          "413": {
            "description": "File too large",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "415": {
            "description": "Unsupported file type",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      },
      "get": {
        "operationId": "listAttachments",
        "summary": "List attachments for expense",
        "tags": [
          "Attachments"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of attachments",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Attachment"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/expenses/{id}/attachments/{aid}": {
      "delete": {
        "operationId": "deleteAttachment",
        "summary": "Delete attachment",
        "tags": [
          "Attachments"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "aid",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Attachment deleted"
          },
          "404": {
            "description": "Attachment not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/test/promote-admin": {
      "post": {
        "operationId": "promoteToAdmin",
        "summary": "Promote a user to admin role",
        "tags": [
          "Test Support"
        ],
        "x-test-only": true,
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "User promoted to admin"
          }
        }
      }
    }
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
...
      "RefreshRequest": {
        "type": "object",
        "description": "Token refresh request",
        "required": [
          "refreshToken"
        ],
        "properties": {
          "refreshToken": {
            "type": "string"
          }
        }
      },
      "AuthTokens": {
        "type": "object",
        "description": "Authentication token pair returned on login or refresh",
        "required": [
          "accessToken",
          "refreshToken",
          "tokenType"
        ],
        "properties": {
          "accessToken": {
            "type": "string"
          },
          "refreshToken": {
            "type": "string"
          },
          "tokenType": {
            "type": "string",
            "example": "Bearer"
          }
        }
      },
...
      "Attachment": {
        "type": "object",
        "description": "File attachment metadata",
        "required": [
          "id",
          "filename",
          "contentType",
          "size",
          "createdAt"
        ],
        "properties": {
          "id": {
            "type": "string"
          },
          "filename": {
            "type": "string"
          },
          "contentType": {
            "type": "string"
          },
          "size": {
            "type": "integer"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
...
      "ErrorResponse": {
        "type": "object",
        "description": "Standardized error response",
        "required": [
          "error"
        ],
        "properties": {
          "error": {
            "type": "string"
          },
          "message": {
            "type": "string"
          },
          "details": {
            "type": "object",
            "additionalProperties": true
          }
        }
      },
...
    }
  }
}